refactor: CORS 설정 클래스 분리
